move DisplayUtilities into UIUtil
